sua them giao dien front crud sanpham admin

+++ routes/web.php
@@ -2,10 +2,10 @@
 
 use Illuminate\Support\Facades\Route;
 
+use App\Http\Controllers\admin\{AdminController, ProductController,};
+
 use App\Http\Controllers\{
     HomeController,
-    AdminController,
-    ProductController,
     DanhmucController,
     AuthController
 };
@@ -23,6 +23,11 @@ use App\Http\Controllers\{
 
 //Frontend
 Route::get('/', [HomeController:: class, 'index']);
+
+Route::get('/sanpham/detail/{id}', [HomeController:: class, 'detail'])->name('detail');
+Route::get('/congiong', [HomeController:: class, 'congiong']);
+Route::get('/search', [HomeController:: class, 'search'])->name('search');
+Route::get('/viewAll', [HomeController:: class, 'viewAll'])->name('viewAll');
 
 Route::get('/login', [AuthController:: class, 'index']);
 Route::post('/login', [AuthController:: class, 'loginPost'])->name('login');
@@ -53,13 +58,13 @@ Route::prefix('/')->middleware('admin.login')->group(function(){
 
 
 //khac
-Route::get('/product', [ProductController:: class, 'index'])->name('product.index');
-Route::get('/product/create', [ProductController:: class, 'create'])->name('product.create');
-Route::post('/product', [ProductController::class, 'store'])->name('product.store');
+Route::get('/admin/product', [ProductController:: class, 'index'])->name('product.index');
+Route::get('/admin/product/create', [ProductController:: class, 'create'])->name('product.create');
+Route::post('/admin/product', [ProductController::class, 'store'])->name('product.store');
 
-Route::get('/product/{product}/edit', [ProductController::class, 'edit'])->name('product.edit');
-Route::put('/product/{product}/update', [ProductController::class, 'update'])->name('product.update');
-Route::delete('/product/{product}/destroy', [ProductController::class, 'destroy'])->name('product.destroy');
+Route::get('/admin/product/edit/{product}', [ProductController::class, 'edit'])->name('product.edit');
+Route::put('/admin/product/update/{product}', [ProductController::class, 'update'])->name('product.update');
+Route::delete('/admin/product/{product}/destroy', [ProductController::class, 'destroy'])->name('product.destroy');
 
 Route::get('/Danhmuc', [DanhmucController::class, 'index'])->name('Danhmuc.index');
